backend/requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.3
blinker==1.7.0
boto3==1.34.74
botocore==1.34.74
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comtypes==1.3.0
defusedxml==0.7.1
diff-match-patch==20230430
distlib==0.3.8
dj-database-url==2.1.0
Django==4.2.9
django-anymail==10.3
django-cors-headers==4.3.1
django-crispy-forms==2.1
django-froala-editor==4.0.13
django-import-export==3.3.7
django-jazzmin==2.6.0
django-storages==1.14.2
django-tinymce==4.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
environs==11.0.0
et-xmlfile==1.1.0
filelock==3.13.1
Flask==3.0.2
idna==3.6
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.3
jmespath==1.0.1
MarkupPy==1.14
MarkupSafe==2.1.5
marshmallow==3.21.1
MouseInfo==0.1.3
numpy==1.26.3
odfpy==1.4.1
opencv-python==4.9.0.80
openpyxl==3.1.2
packaging==24.0
pillow==10.2.0
platformdirs==4.2.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyJWT==2.8.0
PyMsgBox==1.0.9
pyperclip==1.8.2
pypiwin32==223
PyRect==0.2.0
PyScreeze==0.1.30
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyttsx3==2.90
pytweening==1.0.7
pytz==2022.1
pywhatkit==5.4
pywin32==306
PyYAML==6.0.1
requests==2.31.0
rest-framework-simplejwt==0.0.2
s3transfer==0.10.1
setuptools==69.2.0
shortuuid==1.0.13
six==1.16.0
soupsieve==2.5
sqlparse==0.4.2
stripe==8.9.0
tablib==3.5.0
typing_extensions==4.10.0
tzdata==2022.1
uritemplate==4.1.1
urllib3==2.2.1
virtualenv==20.25.0
Werkzeug==3.0.1
wikipedia==1.4.0
xlrd==2.0.1
xlwt==1.3.0
gunicorn